requirements.txt
ipykernel
requests
python-dotenv
discord